refactor testnet cmd into subpackage
