+++ cmd/testnet/down.go
@@ -1,0 +1,25 @@
+package testnet
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"github.com/spf13/cobra"
+)
+
+func DownCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "down",
+		Short: "A convenience command that runs `docker-compose down` on the generated config.",
+		Args:  cobra.NoArgs,
+		RunE: func(_ *cobra.Command, _ []string) error {
+			cmd := []string{"docker-compose", "--file", filepath.Join(generatedConfigDir, "docker-compose.yaml"), "down"}
+			fmt.Println("running:", strings.Join(cmd, " "))
+			if err := replaceCurrentProcess(cmd...); err != nil {
+				return fmt.Errorf("could not run command: %v", err)
+			}
+			return nil
+		},
+	}
+}
