+++ cmd/testnet/utils.go
@@ -1,0 +1,37 @@
+package testnet
+
+import (
+	"os"
+	"os/exec"
+	"syscall"
+)
+
+func replaceCurrentProcess(command ...string) error {
+	if len(command) < 1 {
+		panic("must provide name of executable to run")
+	}
+	executable, err := exec.LookPath(command[0])
+	if err != nil {
+		return err
+	}
+
+	// pass the current environment variables to the command
+	env := os.Environ()
+
+	err = syscall.Exec(executable, command, env)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+type stringSlice []string
+
+func (strings stringSlice) contains(match string) bool {
+	for _, s := range strings {
+		if match == s {
+			return true
+		}
+	}
+	return false
+}
